restriktioner för användaren

+++ Calendar/calendar_event_delete.php
@@ -1,6 +1,6 @@
 <?php
     require_once("../db.php");
-    require_once("../token.php");
+    require_once("../verify_token.php");
     $version = "0.0.8";
     $ok = "OK";
     $error = "Error";
@@ -11,8 +11,12 @@
         $userID = $_GET['userID'];
         $token = $_GET['token'];
     }else{
-        echo json_encode(["Version: "=>$version, "Type: "=>$error, "Data: "=>"You need to log in"]);
+        $json_result = ["Version"=>$version, "Status"=>$ok, "Data"=>"You need to log in"];
+        echo json_encode($json_result);
     }
+    
+    verify_token($userID, $token);
+
 
     if(!empty($_GET['ID'])){
         $ID = $_GET['ID'];
@@ -30,18 +34,33 @@
             $userID = $row['ID'];
             }
     }else {
-        $result = ["Version"=>$version, "Status"=>$ok, "Data"=>"Please log in"];
-        echo json_encode($result);
+        $json_result = ["Version"=>$version, "Status"=>$ok, "Data"=>"Please log in"];
+        echo json_encode($json_result);
     }
 
-    $del = "DELETE FROM calendar_event WHERE ID=?";
+    $del = "DELETE FROM calendar_event WHERE ID=? AND userID=?";
     
     //prepared statement
     $stmt = $conn->prepare($del);
-    $stmt->bind_param("i", $ID);
+    $stmt->bind_param("ii", $ID, $userID);
     $stmt->execute();
-    $result = $stmt->get_result();
 
-    $delete = ["Version"=>$version, "Status"=>$ok, "Data"=>"event removed"];
-    echo json_encode($delete);
+    if ($stmt->affected_rows == 1){
+        $json_result = ["Version"=>$version, "Status"=>$ok, "Data"=>"event removed"];
+        echo json_encode($json_result);
+    }else if ($stmt->affected_rows == 0){
+        $sql = "DELETE FROM calendar_invite WHERE eventID=? AND userID=?";
+
+        $stmt = $conn->prepare($sql);
+        $stmt->bind_param("ii", $ID, $userID);
+        $stmt->execute();
+
+        if ($stmt->affected_rows == 1){
+            $json_result = ["Version"=>$version, "Status"=>$ok, "Data"=>"invite removed"];
+            echo json_encode($json_result);
+        }else{
+            $json_result = ["Version"=>$version, "Status"=>$ok, "Data"=>"You can't delete someone else's event"];
+            echo json_encode($json_result);
+        } 
+    }
 ?>
